Structural material create methods moved and updated descriptions
